# docs/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Little Bear RAG Backend API
  version: 0.1.0
  description: >
    P0 OpenAPI contract for Little Bear RAG backend. The contract covers setup,
    authentication, organization, permission, knowledge base, document import,
    indexing operations, query, audit, healthcheck, config, and internal model
    gateway APIs. P0 uses external embedding, rerank, and LLM provider
    endpoints instead of built-in model simulators, and does not include complete
    config approval or rollback workflows.

    Backend contract paths use /internal/v1. OpenAPI, backend routes, frontend
    SDKs, and contract tests should all use this path prefix consistently.
servers:
  - url: http://localhost:8000
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /internal/v1/knowledge-bases/{kb_id}/documents:
    get:
      tags: [KnowledgeBase]
      summary: List accessible documents
      operationId: listDocuments
      x-token-type: access
      x-required-scope: document:read
      parameters:
        - $ref: "#/components/parameters/KbId"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/Keyword"
        - $ref: "#/components/parameters/Status"
      responses:
        "200":
          description: Document list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentListResponse"
        default:
          $ref: "#/components/responses/Error"
    post:
      tags: [Import]
      summary: Upload files and create import job
      operationId: createUploadDocumentImport
      x-token-type: access
      x-required-scope: document:import
      x-audit-required: true
      parameters:
        - $ref: "#/components/parameters/KbId"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/DocumentUploadRequest"
      responses:
        "202":
          description: Import job queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportJobResponse"
        default:
          $ref: "#/components/responses/Error"
  /internal/v1/knowledge-bases/{kb_id}/document-imports:
    post:
      tags: [Import]
      summary: Create URL or metadata import job
      operationId: createDocumentImport
      x-token-type: access
      x-required-scope: document:import
      x-audit-required: true
      parameters:
        - $ref: "#/components/parameters/KbId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentImportRequest"
      responses:
        "202":
          description: Import job queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportJobResponse"
        default:
          $ref: "#/components/responses/Error"
...
  /internal/v1/import-jobs/{job_id}:
    get:
      tags: [Import]
      summary: Get import job
      operationId: getImportJob
      x-token-type: access
      x-required-scope: import_job:read:self
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        "200":
          description: Import job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportJobResponse"
        default:
          $ref: "#/components/responses/Error"
    patch:
      tags: [Import]
      summary: Update import job status
      operationId: patchImportJob
      x-token-type: access
      x-required-scope: import_job:manage:self
      x-audit-required: true
      parameters:
        - $ref: "#/components/parameters/JobId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportJobPatchRequest"
      responses:
        "200":
          description: Import job updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportJobResponse"
        default:
          $ref: "#/components/responses/Error"
  /internal/v1/import-jobs/{job_id}/retries:
    post:
      tags: [Import]
      summary: Create import job retry
      operationId: createImportJobRetry
      x-token-type: access
      x-required-scope: import_job:manage:self
      x-audit-required: true
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        "202":
          description: Retry job queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportJobResponse"
        default:
          $ref: "#/components/responses/Error"
  /internal/v1/admin/import-jobs:
    get:
      tags: [Import]
      summary: Admin list import jobs
      operationId: adminListImportJobs
      x-token-type: access
      x-required-scope: import_job:read
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/Status"
        - name: stage
          in: query
          schema:
            type: string
        - name: kb_id
          in: query
          schema:
            type: string
        - name: job_type
          in: query
          schema:
            type: string
        - name: created_by
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Import jobs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportJobListResponse"
        default:
          $ref: "#/components/responses/Error"
  /internal/v1/admin/import-jobs/{job_id}:
    get:
      tags: [Import]
      summary: Admin get import job
      operationId: adminGetImportJob
      x-token-type: access
      x-required-scope: import_job:read
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        "200":
          description: Import job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportJobResponse"
        default:
          $ref: "#/components/responses/Error"
  /internal/v1/admin/index-jobs/retries:
    post:
      tags: [AdminIndexOps]
      summary: Retry failed index rebuild jobs
      operationId: adminRetryIndexJobs
      x-token-type: access
      x-required-scope: document:index
      x-audit-required: true
      x-confirmation-required: true
      x-confirmation-condition: always
      parameters:
        - name: x-index-confirm
          in: header
          required: true
          schema:
            type: string
            enum: [retry]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IndexJobRetryRequest"
      responses:
        "202":
          description: Retry jobs queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportJobListResponse"
        default:
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 20
    Keyword:
      name: keyword
      in: query
      schema:
        type: string
    Status:
      name: status
      in: query
      schema:
        type: string
    KbId:
      name: kb_id
      in: path
      required: true
      schema:
        type: string
...
    JobId:
      name: job_id
      in: path
      required: true
      schema:
        type: string
...
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    RequestId:
      type: string
      examples: ["req_001"]
...
    AnyObject:
      type: object
      additionalProperties: true
    Pagination:
      type: object
      required: [page, page_size, total]
      properties:
        page:
          type: integer
        page_size:
          type: integer
        total:
          type: integer
    ErrorResponse:
      type: object
      required: [request_id, error_code, message, retryable]
      properties:
        request_id:
          $ref: "#/components/schemas/RequestId"
        error_code:
          type: string
        message:
          type: string
        stage:
          type: string
        retryable:
          type: boolean
        details:
          $ref: "#/components/schemas/AnyObject"
...
    Visibility:
      type: string
      enum: [department, enterprise]
...
    ImportJobStatus:
      type: string
      enum: [queued, running, retrying, partial_success, success, failed, cancelled]
...
    Document:
      type: object
      required: [id, kb_id, title, lifecycle_status, index_status, owner_department_id, visibility]
      properties:
        id:
          type: string
        kb_id:
          type: string
        folder_id:
          type: [string, "null"]
        title:
          type: string
        lifecycle_status:
          type: string
          enum: [draft, active, archived, deleted]
        index_status:
          type: string
          enum: [none, indexing, indexed, index_failed, blocked]
        owner_department_id:
          type: string
        visibility:
          $ref: "#/components/schemas/Visibility"
        current_version_id:
          type: [string, "null"]
...
    DocumentListResponse:
      type: object
      properties:
        request_id:
          $ref: "#/components/schemas/RequestId"
        data:
          type: array
          items:
            $ref: "#/components/schemas/Document"
        pagination:
          $ref: "#/components/schemas/Pagination"
...
    DocumentUploadRequest:
      type: object
      required: [files]
      properties:
        files:
          type: array
          items:
            type: string
            format: binary
        owner_department_id:
          type: string
        visibility:
          $ref: "#/components/schemas/Visibility"
        folder_id:
          type: [string, "null"]
        idempotency_key:
          type: string
    DocumentImportRequest:
      type: object
      required: [job_type, items]
      properties:
        job_type:
          type: string
          enum: [url, metadata_batch]
        owner_department_id:
          type: string
        visibility:
          $ref: "#/components/schemas/Visibility"
        idempotency_key:
          type: string
        items:
          type: array
          items:
            type: object
            required: [title]
            properties:
              title:
                type: string
              url:
                type: string
                format: uri
              metadata:
                type: object
                additionalProperties: true
    ImportJob:
      type: object
      required: [id, kb_id, status, stage]
      properties:
        id:
          type: string
        kb_id:
          type: [string, "null"]
        job_type:
          type: [string, "null"]
        status:
          $ref: "#/components/schemas/ImportJobStatus"
        stage:
          type: string
          enum: [validate, parse, clean, chunk, embed, index, publish, cleanup, finished]
        document_ids:
          type: array
          items:
            type: string
        error_summary:
          type: [string, "null"]
    ImportJobListItem:
      type: object
      required: [id, kb_id, status, stage, document_count]
      properties:
        id:
          type: string
        kb_id:
          type: [string, "null"]
        job_type:
          type: [string, "null"]
        status:
          $ref: "#/components/schemas/ImportJobStatus"
        stage:
          type: string
          enum: [validate, parse, clean, chunk, embed, index, publish, cleanup, finished]
        document_count:
          type: integer
          minimum: 0
        error_summary:
          type: [string, "null"]
    IndexJobRetryRequest:
      type: object
      required: [job_ids]
      properties:
        job_ids:
          type: array
          minItems: 1
          maxItems: 100
          items:
            type: string
    ImportJobPatchRequest:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [cancelled]
    ImportJobResponse:
      type: object
      required: [request_id, data]
      properties:
        request_id:
          $ref: "#/components/schemas/RequestId"
        data:
          $ref: "#/components/schemas/ImportJob"
    ImportJobListResponse:
      type: object
      properties:
        request_id:
          $ref: "#/components/schemas/RequestId"
        data:
          type: array
          items:
            $ref: "#/components/schemas/ImportJobListItem"
        pagination:
          $ref: "#/components/schemas/Pagination"
...